refactor: photoModal을 파일 내부로 위치 수정
